Refactor VFS v2: Rename tmpfs_v2 module to tmpfs for consistency across the codebase
